move this to api package Fixes #20
