+++ api/export.go
@@ -1,4 +1,5 @@
-package model
+// todo: #20 move this to api package
+package api
 
 type Export interface {
 	Export() (interface{}, error)
